0.2.6

+++ src/routes/accounts.ts
@@ -885,59 +885,70 @@ accountsRouter.delete(`/deletion/start`, async (req: Request, res: Response) => 
       return;
     };
 
+    connection = await dbPool.getConnection();
+    await connection.execute('SET TRANSACTION LEVEL ISOLATION LEVEL SERIALIZABLE;');
+    await connection.beginTransaction();
+
     interface HangoutDetails extends RowDataPacket {
       hangout_id: string,
+      current_step: number,
+      hangout_member_id: number,
     };
 
-    const [hangoutRows] = await dbPool.execute<HangoutDetails[]>(
+    const [hangoutRows] = await connection.execute<HangoutDetails[]>(
       `SELECT
-        hangout_id
+        hangouts.hangout_id,
+        hangouts.current_step,
+        hangout_members.hangout_member_id
       FROM
-        hangout_members
+        hangouts
+      INNER JOIN
+        hangout_members ON hangouts.hangout_id = hangout_members.hangout_id
       WHERE
-        account_id = ? AND
-        is_leader = TRUE;`,
+        hangout_members.account_id = ?;`,
       [accountID]
     );
 
     if (hangoutRows.length > 0) {
-      let hangoutIdsToDelete: string = ``;
+      const hangoutsInVotingStep: HangoutDetails[] = hangoutRows.filter((hangout: HangoutDetails) => hangout.current_step === 3);
 
-      for (let i = 0; i < hangoutRows.length; i++) {
-        if (i + 1 === hangoutRows.length) {
-          hangoutIdsToDelete += `'${hangoutRows[i].hangout_id}'`;
-          continue;
+      if (hangoutsInVotingStep.length > 0) {
+        const hangoutMemberIDs: number[] = hangoutsInVotingStep.map((hangout: HangoutDetails) => hangout.hangout_member_id);
+        const [resultSetHeader] = await connection.execute<ResultSetHeader>(
+          `DELETE FROM
+            votes
+          WHERE
+            hangout_member_id IN (${hangoutMemberIDs.join(', ')})
+          LIMIT ${hangoutMemberIDs.length};`
+        );
+
+        if (resultSetHeader.affectedRows !== hangoutMemberIDs.length) {
+          await connection.rollback();
+          res.status(500).json({ success: false, message: 'Internal server error.' });
+
+          return;
         };
-
-        hangoutIdsToDelete += `'${hangoutRows[i].hangout_id}', `;
       };
 
-      const [resultSetHeader] = await dbPool.execute<ResultSetHeader>(
-        `DELETE FROM
-          hangouts
-        WHERE
-          hangout_id IN (${hangoutIdsToDelete});`
-      );
-
-      if (resultSetHeader.affectedRows === 0) {
-        res.status(500).json({ success: false, message: 'Internal server error.' });
-        return;
-      };
-
-      await dbPool.execute(
+      const hangoutMemberIDs: number[] = hangoutRows.map((hangout: HangoutDetails) => hangout.hangout_member_id);
+      const [resultSetHeader] = await connection.execute<ResultSetHeader>(
         `DELETE FROM
           hangout_members
         WHERE
-          account_id = ?;`,
-        [accountID]
+          hangout_member_id IN (${hangoutMemberIDs.join(', ')})
+        LIMIT ${hangoutMemberIDs.length};`
       );
+
+      if (resultSetHeader.affectedRows !== hangoutMemberIDs.length) {
+        await connection.rollback();
+        res.status(500).json({ success: false, message: 'Internal server error.' });
+
+        return;
+      };
     };
 
     const markedAuthToken: string = `d_${authToken}`;
     const cancellationToken: string = tokenGenerator.generateUniqueToken();
-
-    connection = await dbPool.getConnection();
-    await connection.beginTransaction();
 
     const [resultSetHeader] = await connection.execute<ResultSetHeader>(
       `UPDATE
@@ -969,6 +980,25 @@ accountsRouter.delete(`/deletion/start`, async (req: Request, res: Response) => 
 
     await connection.commit();
     res.status(202).json({ success: true, resData: {} });
+
+
+    const logDescription: string = `${accountDetails.display_name} has left the hangout.`;
+    const currentTimestamp: number = Date.now();
+
+    let logValues: string = '';
+    for (const hangout of hangoutRows) {
+      logValues += `('${hangout.hangout_id}', '${logDescription})', ${currentTimestamp}),`;
+    };
+    logValues.slice(0, -1);
+
+    await dbPool.execute(
+      `INSERT INTO hangout_logs(
+        hangout_id,
+        log_description,
+        log_timestamp
+      )
+      VALUES(${logValues});`
+    );
 
     await sendDeletionEmail(accountDetails.email, accountID, cancellationToken, accountDetails.display_name);
